Added Circle
Map now starts on the coordinates of QUT (this is a placeholder until I get it to snap to the user's location) and creates a red, transparent circle around a pin representing the search area on the map.

Next update will link the size of this ring to the slider at the bottom and on the previous page.